+++ src/mnist.h
@@ -17,6 +17,6 @@ typedef struct {
 
 idx3 read_images_mnist(const char* path);
 idx1 read_labels_mnist(const char* path);
-void get_mnist_image_norm(double* output, idx3* dataset, int index);
+void get_mnist_image_norm(float* output, idx3* dataset, int index);
 void free_mnist_images(idx3* dataset);
 void free_mnist_labels(idx1* dataset);
